initialize tomato formatter
